Rename `CreateDataAppModal`
Changing to `DataAppScaffoldingModal` as `CreateDataAppModal` will include both education and data picker modals
