+++ frontend/src/metabase/writeback/containers/CreateDataAppModal/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./CreateDataAppModal";
+export { default } from "./DataAppScaffoldingModal";
